Cambiado tipo de dato de PhoneNumber a long.

+++ HealthManager/Models/DTO/PatientViewModel.cs
@@ -19,7 +19,7 @@ namespace HealthManager.Models.DTO
         public int Dni { get; set; }
         [Required, Compare("Password"), MinLength(12), RegularExpression("^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*+-]).{12,}$")]
         public string ConfirmPassword { get; set; } = null!;
-        public int? PhoneNumber { get; set; }
+        public long? PhoneNumber { get; set; }
         [Required]
         [MaxLength(1)]
         public string? Gender { get; set; }
